performance updates to reduce resource usage
- reducing resource usage, and introduced control environment for
testing. - security updates to ensure user doesn't have admin priv. -
mac_requirements and linux requirement files separated. - middleware
added to monitor and restrict resource usage from exceeding. - added
additional config to make python env more memory efficient. - lru
handled at redis level instead of handling it in the application level -
file contents downloaded for adding entry instead of storing the
contents locally - removed unwanted variable snipped - reduced
background tasks to reduce the load

+++ backend/scripts/download_model.py
@@ -1,12 +1,15 @@
 import os
 
-from sentence_transformers import SentenceTransformer
-
 model_name = os.environ["EMBEDDING_MODEL_NAME"]
 model_dir = os.path.join(os.path.dirname(__file__), "..", "models", model_name)
 model_dir = os.path.abspath(model_dir)
 
-print(f"Downloading {model_name} to {model_dir} ...")
-model = SentenceTransformer(model_name)
-model.save_pretrained(model_dir)
-print("Model saved.")
+if not os.path.exists(model_dir) or not os.listdir(model_dir):
+    from sentence_transformers import SentenceTransformer
+
+    print(f"Downloading {model_name} to {model_dir} ...")
+    model = SentenceTransformer(model_name)
+    model.save_pretrained(model_dir)
+    print("Model saved.")
+else:
+    print(f"Model already downloaded in {model_dir}.")
